+++ WalletInjector/WalletInjector/Domain/Translator/TradeModelTranslator.swift
@@ -10,8 +10,8 @@ import Foundation
 internal struct TradeModelTranslator: Translator {
     internal func translate(_ entity: TradeEntity) -> TradeModel {
         let unixtime: Int? = entity.unixtime
-        let price: Float? = entity.price
-        let amount: Float? = entity.amount
+        let price: Double? = entity.price
+        let amount: Double? = entity.amount
 
         return TradeModel(unixtime: unixtime, price: price, amount: amount)
     }
